refactor: extract endpoints into separate files
